+++ packages/@expo/cli/src/export/exportStaticAsync.ts
@@ -13,6 +13,7 @@ import prettyBytes from 'pretty-bytes';
 import { inspect } from 'util';
 
 import { getVirtualFaviconAssetsAsync } from './favicon';
+import { persistMetroAssetsAsync } from './persistMetroAssets';
 import { Log } from '../log';
 import { DevServerManager } from '../start/server/DevServerManager';
 import { MetroBundlerDevServer } from '../start/server/metro/MetroBundlerDevServer';
@@ -32,8 +33,9 @@ type Options = {
   outputDir: string;
   minify: boolean;
   exportServer: boolean;
-  basePath: string;
+  baseUrl: string;
   includeMaps: boolean;
+  clear: boolean;
 };
 
 /** @private */
@@ -52,6 +54,7 @@ export async function unstable_exportStaticAsync(projectRoot: string, options: O
     mode: 'production',
     port,
     location: {},
+    resetDevServer: options.clear,
   });
   await devServerManager.startAsync([
     {
@@ -60,6 +63,7 @@ export async function unstable_exportStaticAsync(projectRoot: string, options: O
         port,
         location: {},
         isExporting: true,
+        resetDevServer: options.clear,
       },
     },
   ]);
@@ -112,21 +116,22 @@ export async function getFilesToExportFromServerAsync(
 export async function exportFromServerAsync(
   projectRoot: string,
   devServerManager: DevServerManager,
-  { outputDir, basePath, exportServer, minify, includeMaps }: Options
+  { outputDir, baseUrl, exportServer, minify, includeMaps }: Options
 ): Promise<void> {
   const { exp } = getConfig(projectRoot, { skipSDKVersionRequirement: true });
   const appDir = getRouterDirectoryWithManifest(projectRoot, exp);
 
-  const injectFaviconTag = await getVirtualFaviconAssetsAsync(projectRoot, { outputDir, basePath });
+  const injectFaviconTag = await getVirtualFaviconAssetsAsync(projectRoot, { outputDir, baseUrl });
 
   const devServer = devServerManager.getDefaultDevServer();
   assert(devServer instanceof MetroBundlerDevServer);
 
   const [resources, { manifest, serverManifest, renderAsync }] = await Promise.all([
-    devServer.getStaticResourcesAsync({ mode: 'production', minify, includeMaps }),
+    devServer.getStaticResourcesAsync({ mode: 'production', minify, includeMaps, baseUrl }),
     devServer.getStaticRenderFunctionAsync({
       mode: 'production',
       minify,
+      baseUrl,
     }),
   ]);
 
@@ -142,9 +147,9 @@ export async function exportFromServerAsync(
       const template = await renderAsync(pathname);
       let html = await serializeHtmlWithAssets({
         mode: 'production',
-        resources,
+        resources: resources.artifacts,
         template,
-        basePath,
+        baseUrl,
       });
 
       if (injectFaviconTag) {
@@ -155,12 +160,20 @@ export async function exportFromServerAsync(
     },
   });
 
-  resources.forEach((resource) => {
+  resources.artifacts.forEach((resource) => {
     files.set(
       resource.filename,
       modifyBundlesWithSourceMaps(resource.filename, resource.source, includeMaps)
     );
   });
+
+  if (resources.assets) {
+    await persistMetroAssetsAsync(resources.assets, {
+      platform: 'web',
+      outputDirectory: outputDir,
+      baseUrl,
+    });
+  }
 
   if (exportServer) {
     const apiRoutes = await exportApiRoutesAsync({
@@ -168,6 +181,7 @@ export async function exportFromServerAsync(
       server: devServer,
       appDir,
       manifest: serverManifest,
+      baseUrl,
     });
 
     // Add the api routes to the files to export.
@@ -181,6 +195,7 @@ export async function exportFromServerAsync(
   fs.mkdirSync(path.join(outputDir), { recursive: true });
 
   Log.log('');
+
   Log.log(chalk.bold`Exporting ${files.size} files:`);
   await Promise.all(
     [...files.entries()]
@@ -204,8 +219,9 @@ export function modifyBundlesWithSourceMaps(
   if (filename.endsWith('.js')) {
     // If the bundle ends with source map URLs then update them to point to the correct location.
 
-    // TODO: basePath support
+    // TODO: baseUrl support
     const normalizedFilename = '/' + filename.replace(/^\/+/, '');
+    // Ref: https://developer.chrome.com/blog/sourcemaps/#sourceurl-and-displayname-in-action-eval-and-anonymous-functions
     //# sourceMappingURL=//localhost:8085/index.map?platform=web&dev=false&hot=false&lazy=true&minify=true&resolver.environment=client&transform.environment=client&serializer.output=static
     //# sourceURL=http://localhost:8085/index.bundle//&platform=web&dev=false&hot=false&lazy=true&minify=true&resolver.environment=client&transform.environment=client&serializer.output=static
     return source.replace(/^\/\/# (sourceMappingURL|sourceURL)=.*$/gm, (...props) => {
@@ -232,17 +248,17 @@ export function getHtmlFiles({
 }): string[] {
   const htmlFiles = new Set<string>();
 
-  function traverseScreens(screens: string | { screens: any; path: string }, basePath = '') {
+  function traverseScreens(screens: string | { screens: any; path: string }, baseUrl = '') {
     for (const value of Object.values(screens)) {
       if (typeof value === 'string') {
-        let filePath = basePath + value;
+        let filePath = baseUrl + value;
         if (value === '') {
           filePath =
-            basePath === ''
+            baseUrl === ''
               ? 'index'
-              : basePath.endsWith('/')
-              ? basePath + 'index'
-              : basePath.slice(0, -1);
+              : baseUrl.endsWith('/')
+              ? baseUrl + 'index'
+              : baseUrl.slice(0, -1);
         }
         if (includeGroupVariations) {
           // TODO: Dedupe requests for alias routes.
@@ -251,7 +267,7 @@ export function getHtmlFiles({
           htmlFiles.add(filePath);
         }
       } else if (typeof value === 'object' && value?.screens) {
-        const newPath = basePath + value.path + '/';
+        const newPath = baseUrl + value.path + '/';
         traverseScreens(value.screens, newPath);
       }
     }
@@ -332,12 +348,14 @@ async function exportApiRoutesAsync({
   outputDir,
   server,
   appDir,
+  baseUrl,
   ...props
 }: {
   outputDir: string;
   server: MetroBundlerDevServer;
   appDir: string;
   manifest: ExpoRouterServerManifestV1;
+  baseUrl: string;
 }): Promise<Map<string, string>> {
   const functionsDir = '_expo/functions';
   const funcDir = path.join(outputDir, functionsDir);
@@ -348,6 +366,7 @@ async function exportApiRoutesAsync({
     appDir,
     outputDir: functionsDir,
     prerenderManifest: props.manifest,
+    baseUrl,
   });
 
   Log.log(chalk.bold`Exporting ${files.size} API Routes.`);