+++ packages/babel-preset-expo/src/common.ts
@@ -59,6 +59,10 @@ export function getIsProd(caller: any) {
   return process.env.BABEL_ENV === 'production' || process.env.NODE_ENV === 'production';
 }
 
+export function getBaseUrl(caller: any): string {
+  return caller?.baseUrl ?? '';
+}
+
 export function getIsServer(caller: any) {
   return caller?.isServer ?? false;
 }