+++ packages/expo-asset/src/AssetSourceResolver.native.ts
@@ -1,0 +1,3 @@
+import AssetSourceResolver from 'react-native/Libraries/Image/AssetSourceResolver';
+export default AssetSourceResolver;
+export * from 'react-native/Libraries/Image/AssetSourceResolver';