+++ packages/expo/src/ts-declarations/process.d.ts
@@ -11,6 +11,9 @@ declare const process: {
     EXPO_ROUTER_IMPORT_MODE_WEB?: string;
     EXPO_ROUTER_ABS_APP_ROOT?: string;
     EXPO_ROUTER_APP_ROOT?: string;
+
+    /** Maps to the `experiments.baseUrl` property in the project Expo config. This is injected by `babel-preset-expo` and supports automatic cache invalidation. */
+    EXPO_BASE_URL?: string;
   };
   [key: string]: any;
 };